// src/api/types.ts
export interface ErrorResponse {
  response?: {
    status?: number;
    data?: {
      message?: string;
    };
  };
}